requirements.txt
mkdocs-material~=9.6.0
mkdocs-minify-plugin~=0.8.0
openai-whisper
